divided project into smaller apps
