+++ web_app/adverts/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import Advert
+
+
+admin.site.register(Advert)
