+++ web_app/adverts/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class AdvertsConfig(AppConfig):
+    name = 'adverts'
