+++ web_app/adverts/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+from . import views
+
+app_name = 'adverts'
+urlpatterns = [
+    path('', views.adverts, name='adverts')
+    ]
