+++ web_app/bands/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import Band
+
+
+admin.site.register(Band)
