+++ web_app/bands/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class BandsConfig(AppConfig):
+    name = 'bands'
