+++ web_app/muzi_muzi/admin.py
@@ -1,12 +1,9 @@
 from django.contrib import admin
-from .models import Advert, City, Genre, PrivateMessages, Profession, Role, Users, Band, Videos
+from .models import City, Genre, Videos, Profession, Role
 
-admin.site.register(Advert)
+
 admin.site.register(City)
 admin.site.register(Genre)
-admin.site.register(PrivateMessages)
+admin.site.register(Videos)
 admin.site.register(Profession)
 admin.site.register(Role)
-admin.site.register(Users)
-admin.site.register(Band)
-admin.site.register(Videos)
