+++ web_app/private_messages/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import PrivateMessages
+
+
+admin.site.register(PrivateMessages)
