+++ web_app/private_messages/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class PrivateMessagesConfig(AppConfig):
+    name = 'private_messages'
