+++ web_app/users/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import Users
+
+
+admin.site.register(Users)
