+++ web_app/web_app/urls.py
@@ -18,5 +18,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('muzi-muzi/', include('muzi_muzi.urls'))
+    path('adverts/', include('adverts.urls'))
 ]
